Enhanced table auto-refresh

+++ src/app/shared/table/table.component.ts
@@ -22,7 +22,8 @@ export class TableComponent implements OnInit, AfterViewInit, OnDestroy {
   @Input() dataSource: TableDataSource<any>;
   @ViewChild('searchInput', { static: false }) searchInput: ElementRef;
   public searchPlaceholder = '';
-  public ongoingRefresh = false;
+  private ongoingRefresh = false;
+  public ongoingAutoRefresh = false;
   public sort: MatSort = new MatSort();
   public maxRecords = Constants.INFINITE_RECORDS;
   public numberOfColumns = 0;
@@ -211,7 +212,9 @@ export class TableComponent implements OnInit, AfterViewInit, OnDestroy {
       // Create timer
       this.autoRefeshTimer = setInterval(() => {
         // Reload
-        this.refresh(true);
+        if (!this.ongoingRefresh) {
+          this.refresh(true);
+        }
       }, this.autoRefeshPollingIntervalMillis);
     }
   }
@@ -235,19 +238,21 @@ export class TableComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   public refresh(autoRefresh = false) {
-    // Enable animation in button
-    if (autoRefresh) {
+    if (!this.ongoingRefresh) {
       this.ongoingRefresh = true;
-      this.destroyAutoRefreshTimer();
-    }
-    // Load Data
-    this.dataSource.refreshData(!this.ongoingRefresh).subscribe(() => {
-      // Enable animation in button
       if (autoRefresh) {
-        this.ongoingRefresh = false;
-        this.createAutoRefreshTimer();
+        this.ongoingAutoRefresh = true;
+        this.destroyAutoRefreshTimer();
       }
-    });
+      // Load Data
+      this.dataSource.refreshData(!this.ongoingAutoRefresh).subscribe(() => {
+        this.ongoingRefresh = false;
+        if (autoRefresh) {
+          this.ongoingAutoRefresh = false;
+          this.createAutoRefreshTimer();
+        }
+      });
+    }
   }
 
   public resetFilters() {
